// src/generated/client/models/TokenizerType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Metamind API
 * Brain spec for Metamind.
 *
 * The version of the OpenAPI document: 2.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum TokenizerType {
    WHITESPACE = 'WHITESPACE',
    UNTOKENIZED = 'UNTOKENIZED'
}

export function TokenizerTypeFromJSON(json: any): TokenizerType {
    return TokenizerTypeFromJSONTyped(json, false);
}

export function TokenizerTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): TokenizerType {
    return json as TokenizerType;
}

export function TokenizerTypeToJSON(value?: TokenizerType | null): any {
    return value as any;
}